Require ts-node/register in nyc
Running tests multiple times in a row was producing inconsistent
coverage reports. About half of the time the coverage report was 50%
coverage and the other half of the time the coverage report was at 75%.
In searching the nyc Github repo, I found a Github issue that described
this problem. The issue had a comment by one of the maintainers [1] that
linked to a blog post he'd written that describes a recommended set of
configuration for NYC, Mocha, and TypeScript [2]. There, I noticed that
the example nycrc.json file had configuration to require
'ts-node/register', as mocha does in this repo now. He does not explain
why it's required, and why it should be done in NYC instead of Mocha,
but adding that to our nycrc.json fixed the inconsistent code coverage
reports and they now correctly and consistently report 75%.

[1]: https://github.com/istanbuljs/nyc/issues/537#issuecomment-289637924
[2]: https://azimi.me/2016/09/30/nyc-mocha-typescript.1.html