Update @silvermine/eslint-config and remove eslint dep
The latest version of @silvermine/eslint-config requires a much newer
version of eslint. We don't need the direct dependency on eslint because
grunt-eslint already has a fuzzy dependency on it. Since
@silvermine/eslint-config has a peer dependency requirement that
specifies the minimum version of eslint it requires, the dependency on
eslint from grunt-eslint gets us the correct version.