Change the name of dot files in the template directory
Using the actual name for the dot files (for example `.npmignore`) in
the template creates problems with them being used before the template
is generated. For example, `.npmingnore` will cause all the files listed to be
excluded from the template when publishing to npm. This causes an error
when trying to copy files like `.eslintrc.json`.

This commit renames all the template dot files to use an underscore and
adds the ability to specify a different destination path for template
files.