Fixes #13 Use TS_NODE_FILES flag to enable ambient type defs in tests
ts-node by default runs files in isolation. ts-node analyzes the
dependencies of the file and uses its dependencies to build the type
context in which the file is executed. That means that if a type is not
imported in the file to be executed or in one of its dependencies,
ts-node does not know about it. Because we register the
chai-strictly-equals plugin in the test/setup/before.ts file and nowhere
else, individual test files do not have the chai-strictly-equal types in
its compilation context.

Setting the TS_NODE_FILES flag to `true` tells ts-node to load types
from all of the files available to it per the `tsconfig` configuration,
just as the TypeScript compiler does.