Fixes #12 Split up Gruntfile template and add missing tasks
Split the Gruntfile.js template up into multiple files, one for each
kind of output, so that it's easier to read and maintain. The Gruntfile
had too many `if` statements / branches to be maintainable and easily
understood.

This new organization more easily revealed that some of the outputted
Gruntfiles had missing tasks (`clean` and `watch` for certain outputs).
These missing tasks were added.