Refs #22 Allow passing CLI options to skip prompts
Using the `yeoman-option-or-prompt` library instead of `prompt` allows
us to pass CLI arguments to skip prompting the user for input. This is a
convenient way to generate a project if you are already familiar with
the @silvermine/typescript generator, and it allows us to more easily
add automated tests for the various generated project types. The current
CLI options are:

--projectName {string} - the name of your project

--isBackEnd {boolean} - whether or not your project is meant to be
executed on the back-end (Node.js)

--isLibrary {boolean} - whether or not your project is a library

--isOpenSource {boolean} - whether or not your project will be open
source

--globalVarName {boolean} - (only exists if --isBackend=false and
--isLibrary=true) the name of the global variable your project will
export

Example: `yo @silvermine/typescript --projectName=myProject \
--isBackEnd=true --isLibrary=true --isOpenSource=true`