Fixes #19 Make tslib a prd dependency for libraries
Library projects's `esm` and `commonjs` outputs are meant to be consumed
by other projects. When the TypeScript compiler builds these outputs, it
adds statements to import `tslib` into its outputted files when the
TypeScript compiler's [`importHelpers` setting is `true`][1]. This means
that when the user of the library imports the library's code, the
library's code attempts to import `tslib` and an error is thrown if
`tslib` is not found (because it's a dev dependency and not a production
dependency). Thus, we make `tslib` a production dependency for generated
libraries.

[1]: https://www.typescriptlang.org/docs/handbook/compiler-options.html