fix: Add ci:test command to use --forbid-only flag in CI environment (#40)
We can prevent accidentally merging a commit that uses mocha's `.only`
method to only run a single test or a single `describe` block by passing
the --forbid-only flag when running in a CI environment. To do this, we
define a new "test:ci" npm script. Developers should use `npm test`
locally for development. `npm run test:ci` is used for running tests in
a CI environment.