fix: Use yo's --force-install flag to surface errors in `npm i` (#43)
The `npm install` part of the project generation process was failing but
did not cause the process to exit with an error code. That's Yeoman's
default behavior, unless you supply the `--force-install` flag when
running the generator. We now use that flag when running our tests so
that the tests fail if `npm install` fails.