fix: Ensure all deps are present when `npm install` runs (#43)
The generator runs `npm install`, which triggers npm scripts such as
`prepare`. We've broken the npm installation process up into two parts:
the first part installs all npm dependencies at the specific versions in
package.json, the second part installs a few chosen dependencies using
the `@latest` tag so that generated projects always have the latest
versions. However, because `npm install` was running first and the npm
`prepare` script triggered by `npm install` depends on one of the
dependencies that are installed in the second step, the `prepare` script
was failing.

To fix this, we changed the order of these steps. First, we install the
latest version of those few dependencies, and then run a full `npm
install` so that all dependencies are present when the subsequent npm
scripts run.