chore: upgrade Webpack and webpack-cli to latest versions
With Node.js 17, breaking changes were introduced for libraries that use
hashing algorithms that are not supported in OpenSSL 3 (Webpack 4.x). In
order to future-proof this generator, it's neccessary to move to
Webpack 5.x and upgrade loaders we're using. Fortunately, the config
for Webpack in our generator is minimal. See:
https://nodejs.org/en/blog/release/v17.0.0